+++ kubernetes-server-mock/src/main/java/io/fabric8/kubernetes/client/server/mock/KubernetesCrudDispatcher.java
@@ -17,6 +17,7 @@ package io.fabric8.kubernetes.client.server.mock;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Status;
@@ -203,14 +204,30 @@ public class KubernetesCrudDispatcher extends CrudDispatcher {
           status = removeStatus(source);
         }
 
-        if (contentType != null) {
-          MediaType type = MediaType.parse(contentType);
-          if (!type.subtype().equals(OperationSupport.JSON_PATCH.subtype())) {
+        MediaType mergeType;
+        if (contentType == null) {
+          mergeType = OperationSupport.JSON_PATCH;
+        } else {
+          MediaType mediaType = MediaType.parse(contentType);
+          String subtype = mediaType.subtype();
+
+          if (subtype.equals(OperationSupport.JSON_PATCH.subtype())) {
+            mergeType = OperationSupport.JSON_PATCH;
+          } else if (subtype.equals(OperationSupport.JSON_MERGE_PATCH.subtype())) {
+            mergeType = OperationSupport.JSON_MERGE_PATCH;
+          } else {
             response.setResponseCode(HttpURLConnection.HTTP_UNSUPPORTED_TYPE);
             return response;
           }
         }
-        JsonNode updated = JsonPatch.apply(patch, source);
+
+        JsonNode updated;
+        if (mergeType == OperationSupport.JSON_PATCH)  {
+          updated = JsonPatch.apply(patch, source);
+        } else {
+          ObjectReader objectReader = context.getMapper().readerForUpdating(source);
+          updated = objectReader.readValue(s);
+        }
 
         if (isStatusPath(path)) {
           status = removeStatus(updated);