+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix #3347: OpenShiftConfig missing some configurations from provided Config
 
 #### Improvements
+* Fix #3390: Make mock server to work again with updateStatus by handling JSON_MERGE_PATCH
 * Fix #3316: allow locking deletion to resource version
 * Fix #3350: Replace references to impls with interfaces in several places in the the dsl
 * Fix #3327: Removed generated ResourceHandlers